Update Remove_SCOM_Agent.ps1
Further testing seems to indicate that the script can work under PS v3 as well... at least when testing with "emulating" that version on a newer OS version. Updated the requirements to reflect v3 at minimum. It is known to not work on PS v2 so that will not be tested.